api/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.4
httpx==0.27.0
qdrant-client==1.9.1
sentence-transformers==3.0.1
torch==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
psycopg2-binary==2.9.9
sqlalchemy==2.0.31
redis==5.0.7
python-dotenv==1.0.1
ruff==0.5.0
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0
